# docs/swagger.yaml
basePath: /api/
definitions:
  domain.CreateTaskInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  domain.CreateUserInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  domain.LoginUserInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  domain.Task:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.UpdateTaskInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  http.ErrorResponse:
    properties:
      messages:
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  http.SuccessResponse:
    properties:
      data:
        x-nullable: true
      success:
        example: true
        type: boolean
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for ToDoApp
  title: Golang ToDoApp API
  version: "1.0"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Login user by credentials
      parameters:
      - description: SignIn Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LoginUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Sign In
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Registration user by credentials
      parameters:
      - description: SignUp Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Sign Up
      tags:
      - Auth
  /task:
    get:
      consumes:
      - application/json
      description: Get user tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Task'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiAuth: []
      summary: Getting tasks
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Create task by input data
      parameters:
      - description: input data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiAuth: []
      summary: Creating task
      tags:
      - Task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete task by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiAuth: []
      summary: Deleting task
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: Get one task by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiAuth: []
      summary: Getting one task
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update task by input data
      parameters:
      - description: input data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiAuth: []
      summary: Updating task
      tags:
      - Task
securityDefinitions:
  ApiAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"